Moved Coord to separate module.
